use LayerCompressorMixin
